archive/requirements.txt
requests==2.32.3
pandas==2.2.3
beautifulsoup4==4.12.3
flask==3.1.3
Jinja2>=3.1.2
psycopg2-binary==2.9.9
